+++ src/hooks/useStripe.tsx
@@ -48,6 +48,7 @@ import {
   verifyMicrodepositsForPayment,
   verifyMicrodepositsForSetup,
   canAddCardToWallet,
+  getIssuingApiVersion,
   collectBankAccountToken,
   collectFinancialConnectionsAccounts,
   resetPaymentSheetCustomer,
@@ -224,6 +225,10 @@ export function useStripe() {
     []
   );
 
+  const _getIssuingApiVersion = useCallback(async (): Promise<string> => {
+    return getIssuingApiVersion();
+  }, []);
+
   const _collectBankAccountToken = useCallback(
     async (clientSecret: string): Promise<FinancialConnections.TokenResult> => {
       return collectBankAccountToken(clientSecret);
@@ -333,6 +338,7 @@ export function useStripe() {
     verifyMicrodepositsForPayment: _verifyMicrodepositsForPayment,
     verifyMicrodepositsForSetup: _verifyMicrodepositsForSetup,
     canAddCardToWallet: _canAddCardToWallet,
+    getIssuingApiVersion: _getIssuingApiVersion,
     collectBankAccountToken: _collectBankAccountToken,
     collectFinancialConnectionsAccounts: _collectFinancialConnectionsAccounts,
     /**